Fix-Link-And-TableColumn-And-Get-Required-Data
