+++ src/arkfbp/index.ts
@@ -1,8 +1,7 @@
 import { runWorkflowByClass } from 'arkfbp/lib/flow'
-import { debug } from 'console'
 import { Loading } from 'element-ui'
-import { url } from 'inspector'
 import { FlowInput } from './flow'
+import Filter from '@/utils/filter'
 
 const flows = require.context('@/flows', true, /\.ts$/)
 
@@ -13,19 +12,6 @@ function getUrl(url: string, data: any) {
 
   const property = url.slice(url.indexOf('<') + 1, url.lastIndexOf('>'))
   return url.slice(0, url.indexOf('<')) + data[property] + url.slice(url.indexOf('>') + 1)
-}
-
-function toProcessData(v: any, vd: any): number {
-  const m = /=(\S*)]/
-  const fv = v.match(m)[1]
-  const fvd = vd
-  let outcome = 0
-  fvd.items.forEach((vditem: any, vdindex: number) => {
-    if (vditem.prop === fv) {
-      outcome = vdindex
-    }
-  })
-  return outcome
 }
 
 export function runFlowByFile(filename: string, inputs: FlowInput) {
@@ -39,7 +25,6 @@ export async function runAction(action: {flow: string, inputs: FlowInput}) {
     const loading = Loading.service({ fullscreen: true })
     const outputs = await runFlowByFile(action.flow, action.inputs)
     loading.close()
-
     return outputs
   }
 }
@@ -73,7 +58,7 @@ export async function runFlow(state: any, flow: any, data: any) {
         const vs = flow.request[key].split('.')
         vs.forEach((v: string) => {
           if (v.slice(0, 11) === 'items[prop=') {
-            const res = toProcessData(v, temp)
+            const res = Filter(v, temp)
             temp = temp['items'][res]
           } else {
             temp = temp[v]
