+++ src/config/json/testpage.json
@@ -173,8 +173,8 @@
   },
   "pagination": {
     "currentPage": 1,
-    "pageCount": 20,
-    "total": 100,
+    "pageCount": 10,
+    "total": 0,
     "background": "",
     "action": "toFilter"
   },
