+++ src/flows/retrieve/nodes/client.ts
@@ -1,32 +1,30 @@
 import { FunctionNode } from 'arkfbp/lib/functionNode'
-
+import Filter from '@/utils/filter'
 export class Client extends FunctionNode {
   async run() {
-
     const state = this.$state.fetch()
-    
     if (state.clientServer == null) {
       return this.inputs
     }
 
     Object.keys(state.clientServer).forEach((key) => {
       const ks = key.split('.')
-      
       let tempC = state.client
-      
       for (let i = 0; i < ks.length - 1; i++) {
-        tempC = tempC[ks[i]]
+        if (ks[i].slice(0, 11) === 'items[prop=') {
+          const res = Filter(ks[i], tempC)
+          tempC = tempC['items'][res]
+        } else {
+          tempC = tempC[ks[i]]
+        }
       }
-      
       const vs = state.clientServer[key].split('.')
-      
       let tempS = this.inputs
       for (let i = 0; i < vs.length - 1; i++) {
         tempS = tempS[vs[i]]
       }
       tempC[ks[ks.length - 1]] = tempS[vs[vs.length - 1]]
     })
-    
     return this.inputs
   }
 }
